Move tip-sheets language select into seperate comp
