+++ src/app/tips/tips.component.html
@@ -7,6 +7,7 @@
             <div class="nav-card-item-content">
               <p>Our open-source and evidence-based parenting tips and resources are available in more than 100 languages!</p>
               <p> Select your preferred language, and see the parenting tip sheets, as well as other parenting materials, available in the language of your choice:</p>
+              <app-tips-lang-select [languages]="languages" (onLanguageChange)="onLanguageChange($event)"></app-tips-lang-select>
               <app-tip-sheets [currentLanguage]="tipSheetLang"></app-tip-sheets>
               <p>If you can't find a parenting tip sheet and/or other parenting resources available in your language, please send us an email at: <a href="mailto:translate@covid19parenting.com">translate@covid19parenting.com</a> </p>
               <p>Please note that these translations were not created by the World Health Organization (WHO), and WHO is not responsible for the content or accuracy of these translations.</p>
